requirements.txt
djangorestframework==3.14.0
pytest-django==4.5.2
Django==4.2
